Refactoring structure
Moving *Service to a separate package and renaming to *Operations
