+++ src/main/java/org/trello4j/Trello.java
@@ -1,13 +1,21 @@
 package org.trello4j;
 
+import org.trello4j.core.ActionOperations;
+import org.trello4j.core.BoardOperations;
+import org.trello4j.core.CardOperations;
+import org.trello4j.core.ChecklistOperations;
+import org.trello4j.core.ListOperations;
+import org.trello4j.core.MemberOperations;
+import org.trello4j.core.NotificationOperations;
+import org.trello4j.core.OrganizationOperations;
 import org.trello4j.model.Type;
 
 /**
  * The Interface Trello.
  */
-public interface Trello extends OrganizationService, NotificationService,
-		BoardService, CardService, ActionService, ListService, MemberService,
-		ChecklistService, TokenService {
+public interface Trello extends OrganizationOperations, NotificationOperations,
+		BoardOperations, CardOperations, ActionOperations, ListOperations, MemberOperations,
+		ChecklistOperations, TokenService {
 
 	/**
 	 * Gets the type.
